+++ FireworkRNSDKSample/src/screens/Shopping.tsx
@@ -92,6 +92,8 @@ function Shopping() {
               playerStyle: 'full',
               videoCompleteAction: 'advanceToNext',
               showShareButton: true,
+              showMuteButton: true,
+              showPlaybackButton: true,
             }}
             onVideoFeedLoadFinished={(error?: FWError) => {
               console.log('onVideoFeedLoadFinished error', error);