+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # CHANGELOG
 
+## [2.0.0]
+
+### Added
+
+- Add `getShareBaseURL` method in `FireworkSDK` class
+- Add `getAdBadgeConfiguration` method in `FireworkSDK` class(Only supported on iOS)
+- Add `setAdBadgeConfiguration` method in `FireworkSDK` class(Only supported on iOS)
+- Add `getVideoLaunchBehavior` method in `FireworkSDK` class
+- Add the ability to programmatically start or stop the floating player on the Android side
+- Support passing `shareBaseURL` and `videoLaunchBehavior` in `FireworkSDK` class `init` method
+
+### Changed
+
+- Upgrade Firework Android SDK from V5 to V6
+
+### Breaking Changes
+
+- Remove support for story block on the Android side
+- Remove support for playlist group video feed source on the Android side
+- Remove support for app-level language setting
+- Change the parameter and return value in `FireworkSDK` class `init` method
+- Remove support for custom layout name on the Android side
+- Remove `customCTALinkContentPageRouteName` property in `FireworkSDK` class
+- Remove set and get accessors of `shareBaseURL` property in `FireworkSDK` class
+- Remove set and get accessors of `adBadgeConfiguration` property in `FireworkSDK` class
+- Remove `appComponentName` property in `FireworkSDK` class
+- Remove `pushNativeContainer` method in `FWNavigator` class
+- Remove support for `onLiveStreamEvent` callback in `FWNavigator` class on the Android side
+- Remove support for `onLiveStreamChatEvent` callback in `FWNavigator` class on the Android side
+- Remove `onClickCartIcon` callback in `VideoShopping` class
+- Remove `onWillDisplayProduct` callback in `VideoShopping` class
+- Remove `NewNativeContainerProps` type
+- Remove `ClickCartIconCallback` type
+- Remvoe `WillDisplayProductCallback` type
+- Remove `WillDisplayProductEvent` interface
+
 ## [1.8.0]
 
 ### Added